# src/Presentation.Web.Server/wwwroot/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "v1"
  },
  "paths": {
...
    "/api/tenants/{tenantId}/catalog/publishers/{id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogPublisher",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/catalog/publishers/{id}/books": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogPublisherBooks",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/catalog/publishers": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetCatalogPublishers",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublisherModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/inventory/stocks/{id}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "GetInventoryStock",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/inventory/stocks": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "GetInventoryStocks",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "operationId": "CreateInventoryStock",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/StockModel"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModel"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenantId}/inventory/stocks/{id}/movements": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "operationId": "CreateInventoryStockMovement",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/StockMovementModel"
                  }
                ]
              }
            }
          },
          "x-position": 3
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModel"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TagModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Edition": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "Sku": {
            "type": "string",
            "nullable": true
          },
          "Isbn": {
            "type": "string",
            "nullable": true
          },
          "Price": {
            "type": "number",
            "format": "decimal"
          },
          "Publisher": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/BookPublisherModel"
              }
            ]
          },
          "PublishedDate": {
            "type": "string",
            "format": "date"
          },
          "Keywords": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "Authors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookAuthorModel"
            }
          },
          "Categories": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookCategoryModel"
            }
          },
          "Chapters": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BookChapterModel"
            }
          },
          "Tags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TagModel"
            }
          },
          "Version": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookPublisherModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookAuthorModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BookCategoryModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BookChapterModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Number": {
            "type": "integer",
            "format": "int32"
          },
          "Content": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PublisherModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Description": {
            "type": "string",
            "nullable": true
          },
          "Version": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StockModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "TenantId": {
            "type": "string",
            "nullable": true
          },
          "Sku": {
            "type": "string",
            "nullable": true
          },
          "QuantityOnHand": {
            "type": "integer",
            "format": "int32"
          },
          "QuantityReserved": {
            "type": "integer",
            "format": "int32"
          },
          "ReorderThreshold": {
            "type": "integer",
            "format": "int32"
          },
          "ReorderQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "UnitCost": {
            "type": "number",
            "format": "decimal"
          },
          "Location": {
            "type": "string",
            "nullable": true
          },
          "LastRestockedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Movements": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/StockMovementModel"
            }
          },
          "Adjustments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/StockAdjustmentModel"
            }
          }
        }
      },
      "StockMovementModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Quantity": {
            "type": "integer",
            "format": "int32"
          },
          "Type": {
            "type": "integer",
            "format": "int32"
          },
          "Reason": {
            "type": "string",
            "nullable": true
          },
          "Timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StockAdjustmentModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "QuantityChange": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "OldUnitCost": {
            "type": "number",
            "format": "decimal"
          },
          "NewUnitCost": {
            "type": "number",
            "format": "decimal"
          },
          "Reason": {
            "type": "string",
            "nullable": true
          },
          "Timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "oauth2",
        "description": "Oidc Authentication",
        "flows": {
          "implicit": {
            "authorizationUrl": "/protocol/openid-connect/auth",
            "tokenUrl": "/protocol/openid-connect/token"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}